Fix dedenting binary data.
Without this patch, `typedruby` panics on the test case:

```ruby
def hi
  <<~END.strip
    \x80
  END
end

```

While allowing the (semantically identical)

```ruby
def hi
  "\x80"
end
```

Constructing invalid utf-8 `String`s is likely to eventually lead to
sadness, so this suggests that a fuller fix is going to require
representing String literals as a `Vec<u8>`, maybe with an encoding
tag, instead. But that bug already exists in the latter test case, so
this commit should fix a panic without making the situation
fundamentally worse.