Fix a memory overflow in the parser.
We weren't using a `std::vector` as our stack memory, but were not
properly managing `size` vs `capacity`, or resizing it as-needed.

From a read of the lexer source, large nested string interpolation is
probably the only case that can trigger this, which explains why the
default stack size of 16 probably sufficed to date.